Create Insertion_at_certain_pos SL
Insert a node at a specific position in a linked list
Last Updated : 16 Sep, 2024
Given a singly linked list, a position pos, and data, the task is to insert that data into a linked list at the given position. 

Examples:

Input: 3->5->8->10, data = 2, pos = 2
Output: 3->2->5->8->10


Input: 3->5->8->10, data = 11, pos = 5
Output: 3->5->8->10->11 



https://media.geeksforgeeks.org/wp-content/uploads/20240902143425/Insertion-at-a-Specific-Position-of-the-Singly-Linked-List-copy.webp


[Expected Approach] Using Iterative Method – O(n) time and O(1) space:
- Initialize a variable , say curr points to head and allocate the memory to the new node with the given data.
- Traverse the Linked list using curr pointer upto position-1 nodes.
If curr’s next is not null , then next pointer of the new node points to the next of curr node.
- The next pointer of current node points to the new node.
return the head of linked list.