Create Single Linked List (Deleting the First Node)
Original list: 30 -> 20 -> 10 -> NULL
After deleting the first node: 20 -> 10 -> NULL

Check if the list is empty. If it is, there’s nothing to delete.
If the list is not empty:
Save the current head node in a temporary pointer.
Move the head pointer to the next node.
Delete the saved node to free memory.