Create insertion at beg(Doubly _LIinkedlist)
Create the New Node: First, you need to create a new node. This involves allocating memory for the node and initializing its data field with the value you want to insert. The prev and next pointers of the new node should initially be set to nullptr.

Handle Empty List: Check if the list is currently empty (i.e., the head pointer is nullptr). If it is, the new node becomes the head of the list. Its prev and next pointers will remain nullptr in this case.

Update New Node's next Pointer: If the list is not empty, the next pointer of the new node should be set to point to the current head of the list. This establishes the link between the new node and the existing list.

Update Current Head's prev Pointer: The prev pointer of the current head node must be updated to point to the new node. This is the crucial step that creates the "doubly linked" connection, allowing traversal in both directions.

Update the Head Pointer: Finally, the head pointer of the list itself must be updated to point to the newly created node. This makes the new node the new beginning of the list.